+++ app/src/main/java/com/sdzshn3/android/news247/ViewModel/ScienceViewModel.java
@@ -12,6 +12,7 @@ import retrofit2.Callback;
 import retrofit2.Response;
 
 import com.sdzshn3.android.news247.Fragments.ScienceNewsFragment;
+import com.sdzshn3.android.news247.Repositories.ScienceRepository;
 import com.sdzshn3.android.news247.Retrofit.ApiService;
 import com.sdzshn3.android.news247.Retrofit.Article;
 import com.sdzshn3.android.news247.Retrofit.Client;
@@ -20,53 +21,18 @@ import com.sdzshn3.android.news247.Retrofit.NewsModel;
 import java.util.List;
 
 public class ScienceViewModel extends AndroidViewModel {
-    private static MutableLiveData<List<Article>> data = new MutableLiveData<>();
-    private ApiService apiService;
-    private static Call<NewsModel> call;
+    private ScienceRepository repository;
 
     public ScienceViewModel(@NonNull Application application) {
         super(application);
-        apiService = Client.getApiService();
-        loadData();
-    }
-
-    private void loadData() {
-        call = apiService.getResponse(ScienceNewsFragment.URL);
-        call.enqueue(new Callback<NewsModel>() {
-            @Override
-            public void onResponse(Call<NewsModel> call, Response<NewsModel> response) {
-                if (response.isSuccessful()) {
-                    data.postValue(response.body().getArticles());
-                }
-            }
-
-            @Override
-            public void onFailure(Call<NewsModel> call, Throwable t) {
-                Log.e("ScienceViewModel", "onFailure", t);
-                data.postValue(null);
-            }
-        });
+        repository = new ScienceRepository();
     }
 
     public void refresh() {
-        call = apiService.getResponse(ScienceNewsFragment.URL);
-        call.clone().enqueue(new Callback<NewsModel>() {
-            @Override
-            public void onResponse(Call<NewsModel> call, Response<NewsModel> response) {
-                if (response.isSuccessful()) {
-                    data.postValue(response.body().getArticles());
-                }
-            }
-
-            @Override
-            public void onFailure(Call<NewsModel> call, Throwable t) {
-                Log.e("ScienceViewModel", "onFailure", t);
-                data.postValue(null);
-            }
-        });
+        repository.refresh();
     }
 
     public LiveData<List<Article>> getData() {
-        return data;
+        return repository.getData();
     }
 }
